Make tiles configurable via a config file.
This commit kills the entire tiletrans -> gametiles.bin mechanism.
Instead, a new configuration file (vultures_tiles.conf) is read on startup;
the tiles are all saved as individual png files in the direcory tiles/

This should allow for easier artist participation since viewing a new tile ingame
will only require a change of one line in the configfile and a restart, rather than a recompile.

This code is (at best) beta quality:
 - the code needs commenting, possibly rewriting for readability in places, ...
 - the new configuration file nees documenting with comments, possibly some format description
 - there are known bugs (and very likely some unknown ones, too). For example for some
   yet-to-be-determined reason the slime mold isn't drawn with it's tile
   It should however be play(-test-)able
 - Only the unix makefiles for nethack/slashem have been updated; windows (and probably mac?) are TODO
 - the new code needs to be run under valgrind; it definitely leaks