split nh window and vultures windows
not every vultures window needs to be a nh window, in fact only a few
toplevel windows are. On the other hand Nethack requests a new nh window
before we even know whether it will be an inventory or a menu window, so
we don't want to create a vultures window immediately.
Splitting the 2 from each other makes it possible to kill the ugly
"replace_window" code that was used to convert menus to inventories